# aarch64: Fix vector register handling and apply optimizations