--- a/aarch64.c
+++ b/aarch64.c
@@ -22,32 +22,31 @@ __asm__(
   ".type co_swap #function\n"
   "co_swap:\n"
 
-  "stp d8,d9,[x1]\n"
-  "stp d10,d11,[x1,#16]\n"
-  "stp d12,d13,[x1,#32]\n"
-  "stp d14,d15,[x1,#48]\n"
-  "str x19,[x1,#72]\n"
-  "stp x20,x21,[x1,#80]\n"
-  "stp x22,x23,[x1,#96]\n"
-  "stp x24,x25,[x1,#112]\n"
-  "stp x26,x27,[x1,#128]\n"
-  "stp x28,x29,[x1,#144]\n"
   "mov x16,sp\n"
-  "stp x16,x30,[x1,#160]\n"
-
-  "ldp d8,d9,[x0]\n"
-  "ldp d10,d11,[x0,#16]\n"
-  "ldp d12,d13,[x0,#32]\n"
-  "ldp d14,d15,[x0,#48]\n"
-  "ldr x19,[x0,#72]\n"
-  "ldp x20,x21,[x0,#80]\n"
-  "ldp x22,x23,[x0,#96]\n"
-  "ldp x24,x25,[x0,#112]\n"
-  "ldp x26,x27,[x0,#128]\n"
-  "ldp x28,x29,[x0,#144]\n"
-  "ldp x16,x17,[x0,#160]\n"
+  "stp x16,x30,[x1]\n"
+  "ldp x16,x30,[x0]\n"
   "mov sp,x16\n"
-  "br x17\n"
+  "stp x19,x20,[x1,#16]\n"
+  "ldp x19,x20,[x0,#16]\n"
+  "stp x21,x22,[x1,#32]\n"
+  "ldp x21,x22,[x0,#32]\n"
+  "stp x23,x24,[x1,#48]\n"
+  "ldp x23,x24,[x0,#48]\n"
+  "stp x25,x26,[x1,#64]\n"
+  "ldp x25,x26,[x0,#64]\n"
+  "stp x27,x28,[x1,#80]\n"
+  "ldp x27,x28,[x0,#80]\n"
+  "str x29,[x1,#96]\n"
+  "ldr x29,[x0,#96]\n"
+  "stp d8,d9,[x1,#112]\n"
+  "ldp d8,d9,[x0,#112]\n"
+  "stp d10,d11,[x1,#128]\n"
+  "ldp d10,d11,[x0,#128]\n"
+  "stp d12,d13,[x1,#144]\n"
+  "ldp d12,d13,[x0,#144]\n"
+  "stp d14,d15,[x1,#160]\n"
+  "ldp d14,d15,[x0,#160]\n"
+  "br x30\n"
 
   ".size co_swap, .-co_swap\n"
 );
@@ -64,9 +63,9 @@ cothread_t co_derive(void* memory, unsigned int size, void (*entrypoint)(void)) 
   if(handle = (unsigned long*)memory) {
     unsigned int offset = (size & ~15);
     unsigned long* p = (unsigned long*)((unsigned char*)handle + offset);
-    handle[19] = (unsigned long)p;           /* x29 (frame pointer) */
-    handle[20] = (unsigned long)p;           /* x30 (stack pointer) */
-    handle[21] = (unsigned long)entrypoint;  /* x31 (link register) */
+    handle[0]  = (unsigned long)p;           /* x16 (stack pointer) */
+    handle[1]  = (unsigned long)entrypoint;  /* x30 (link register) */
+    handle[12] = (unsigned long)p;           /* x29 (frame pointer) */
   }
 
   return handle;