@xnox 2020-07-01
@mwhudson could you please review this? This would speed up LXD on AWS.

@Sintendo BTW can we use atomics as available on Gravitron2 for any of this? When available / detected with hwcaps or an ifunc?