@Sintendo 2020-07-02
> BTW can we use atomics as available on Gravitron2 for any of this? When available / detected with hwcaps or an ifunc?

I'm assuming you're talking about using SWP instructions or something? Unfortunately SWP loads from and writes to a single memory location, while we need to write to one memory location while loading from another.